==> 00_experiment_with_erddap.ipynb <==
from erddapy import ERDDAP
import pandas as pd
import json




# load a list of all the ERDDAPs and their urls.
# This is saved as a JSON with the code bc I cant get the direct link to the git site to work.
f = open('erddaps.json')
data_json = pd.read_json(f)
erddap_links = data_json['url'].unique()

erddap_links=erddap_links

erddap_links
# # load in all possible erddap servers straight from site (not working...why??):
# url = 'https://github.com/IrishMarineInstitute/awesome-erddap/blob/df70c66a0784a384f398492d356df46dae4281d4/erddaps.json'
# response = urllib.request.urlopen(url)
# data_json = json.loads(response.read())

# set up for the ERDDAP API
e = ERDDAP(
    server="https://erddap.bco-dmo.org/erddap", # pick a specific ERDDAP server
    protocol="tabledap", # pick if you want grid(raster) or table data
    response="csv") # pick type of data output
e.dataset_id = "bcodmo_dataset_817952" # define the dataset on that ERDDAP server

# define by space/time/variable
e.constraints = {
    "time>=": "2017-01-13T00:00:00Z",
    "time<=": "2017-01-16T23:59:59Z"}


# look at the station metadata for this query:
info_url = e.get_info_url()
pd.read_csv(info_url)# make a dataframe for all the metadata for this station


# define the variables you want to download (or leave blank and download all columns)
e.variables = [
    "longitude",
    "latitude",
    "time",
    "Temperature"]

# or load the data into a dataframe
url = e.get_download_url()
df_bcodmo = e.to_pandas(parse_dates=True)
